==> Heart_Attack/01_01_ED_patients_to_hosp.ipynb <==
"""
Καταγράφω τους ασθενείς που πέρασαν από τα
επείγοντα με έμφραγμα και κατέληξαν στο νοσοκομείο.

949.172 -> Συνολικές εγγραφές στον πίνακα των διαγνώσεων.
448.972 -> Εγγραφές που μας δείχνουν τους ασθενείς που νοσηλεύτηκαν.
  2.292 -> Είναι οι διαγνώσεις εγκεφαλικού.
  2.233 -> Εισαγωγές με εγκεφαλικό νοσηλεύτηκαν
  2.145 -> Μοναδικοί ασθενείς που διαγνώστηκαν με εγκεφαλικό
  2.094 -> Μοναδικοί ασθενείς με εγκεφαλικό που νοσηλεύτηκαν.
     59 -> Εισαγωγές με εγκεφαλικό που δεν νοσηλεύτηκαν.
     59 -> Μοναδικοί ασθενείς με εγκεφαλικό που δεν νοσηλεύτηκαν.
"""

import pandas as pd

df_patients = pd.read_csv('D:\Thesis\mimic-iv-1.0\ed\diagnosis.csv.gz')
df_edstays = pd.read_csv('D:\Thesis\mimic-iv-1.0\ed\edstays.csv.gz')

#Ασθενείς με εγκεφαλικό.
ed_cerebral_infarction_diagnoses = df_patients[
    (df_patients['icd_code'] =="41000") | (df_patients['icd_code'] =="41001") |
    (df_patients['icd_code'] =="41002") | (df_patients['icd_code'] =="41010") |
    (df_patients['icd_code'] =="41011") | (df_patients['icd_code'] =="41012") |
    (df_patients['icd_code'] =="41020") | (df_patients['icd_code'] =="41021") |
    (df_patients['icd_code'] =="41022") | (df_patients['icd_code'] =="41030") |
    (df_patients['icd_code'] =="41031") | (df_patients['icd_code'] =="41032") |
    (df_patients['icd_code'] =="41040") | (df_patients['icd_code'] =="41041") |
    (df_patients['icd_code'] =="41042") | (df_patients['icd_code'] =="41050") |
    (df_patients['icd_code'] =="41051") | (df_patients['icd_code'] =="41052") |
    (df_patients['icd_code'] =="41060") | (df_patients['icd_code'] =="41061") |
    (df_patients['icd_code'] =="41062") | (df_patients['icd_code'] =="41070") |
    (df_patients['icd_code'] =="41071") | (df_patients['icd_code'] =="41072") |
    (df_patients['icd_code'] =="41080") | (df_patients['icd_code'] =="41081") |
    (df_patients['icd_code'] =="41082") | (df_patients['icd_code'] =="41090") |
    (df_patients['icd_code'] =="41091") | (df_patients['icd_code'] =="41092") |
    (df_patients['icd_code'] =="I2109") | (df_patients['icd_code'] =="I2119") |
    (df_patients['icd_code'] =="I2111") | (df_patients['icd_code'] =="I2129") |
    (df_patients['icd_code'] =="I213") | (df_patients['icd_code'] =="I214") |
    (df_patients['icd_code'] =="I219") | (df_patients['icd_code'] =="I21A1") |
    (df_patients['icd_code'] =="I21A9") | (df_patients['icd_code'] =="I21") |
    (df_patients['icd_code'] =="I210") | (df_patients['icd_code'] =="I2101") |
    (df_patients['icd_code'] =="I2102") | (df_patients['icd_code'] =="I211") |
    (df_patients['icd_code'] =="I212") | (df_patients['icd_code'] =="I2121") |
    (df_patients['icd_code'] =="I21A") | (df_patients['icd_code'] =="I22") |
    (df_patients['icd_code'] =="I220") | (df_patients['icd_code'] =="I221") |
    (df_patients['icd_code'] =="I222") | (df_patients['icd_code'] =="I228") |
    (df_patients['icd_code'] =="I229") | (df_patients['icd_code'] =="I23") |
    (df_patients['icd_code'] =="I230") | (df_patients['icd_code'] =="I231") |
    (df_patients['icd_code'] =="I232") | (df_patients['icd_code'] =="I233") |
    (df_patients['icd_code'] =="I234") | (df_patients['icd_code'] =="I235") |
    (df_patients['icd_code'] =="I236") | (df_patients['icd_code'] =="I238") |
    (df_patients['icd_code'] =="42979")
]

# Διαγνώσεις εγκεφαλικού (2292).
ed_cerebral_infarction_diagnoses

"""
Ενοποίηση των στοιχείων έτσι ώστε να κρατήσουμε τους
ασθενείς που έχουν hadm_id πράγμα που σημαίνει ότι
νοσηλεύτηκαν στο νοσοκομείο. Από εκεί θα δούμε αν
κατέληξαν στις ΜΕΘ. Με βάση την βιβλιογραφία της MimicIV
το hadm_id δεν δείχνει την άμεση μετάβαση στις ΜΕΘ.
Βέβαια το ίδιο hadm_id ακολουθεί τον ασθενή καθόλη την 
διάρκεια της εισαγωγής αλλά θα κρατήσω της βιβλιογραφίας.
"""
rslt_df = pd.merge(df_edstays, ed_cerebral_infarction_diagnoses, on=['subject_id', 'stay_id'] )

rslt_df

"""
Έλεγχω αν υπάρχουν nan κελιά, έτσι μπορώ
να δω αν κάποιος ασθενής πέρασε από τα 
επείγοντα στο νοσοκομείο ή όχι. Οι κενές
περιοχές δηλώνουν ότι δεν έφυγε χωρίς εισαγωγή.
"""
rslt_df.isna().sum()

# Κρατώ μόνο τις γραμμές που έχουν τιμές στο hadm_id
rslt_df = rslt_df[rslt_df['hadm_id'].notna()]

rslt_df.to_csv('D:\Thesis\exports\o1_ed_patients_to_hosp.csv', index=False)

# Μοναδικοί ασθενείς που διαγνώστηκαν με εγκεφαλικό (2145)
ed_cerebral_infarction_diagnoses.drop_duplicates(subset=['subject_id']).info()

# Ελέγχω εκ νέου για κενές τιμές στο hadm
rslt_df[rslt_df['hadm_id'].isna()].drop_duplicates(subset=['subject_id']).info()

# Εισαγωγές με εγκεφαλικό νοσηλεύτηκαν (2233)
rslt_df.info()

rslt_df.drop_duplicates(subset=['subject_id']).info()
