==> Heart_Attack/01_02_ED_mean_vitals.ipynb <==
"""
Ενσωματώνω τα στοιχεία των ζωτικών ενδείξεων
ομαδοποιημένα ανά 8ώρο στους ασθενείς με
εγκεφαλικό που πέρασαν από τα επείγοντα στο
νοσοκομείο.

1.651.119 -> Συνολικές εγγραφές στο vitalsigns.
  752.146 -> Συνολικές κατά την 8ώρη ομαδοποίηση.
    2.645 -> Εγγραφές 8ώρου για τους ασθενείς που
             νοσηλεύτηκαν με εγκεφαλικό από τα
             επείγοντα.
  
"""

import pandas as pd

#Εισαγωγή dataframes
df_vitalsign = pd.read_csv('D:\Thesis\mimic-iv-1.0\ed\Vitalsign.csv.gz')
df_ed_patients = pd.read_csv('D:\Thesis\exports\o1_ed_patients_to_hosp.csv',
                             usecols = ['subject_id', 'hadm_id', 'stay_id'])

# Μετατροπή της στήλης charttime από object σε datetime
df_vitalsign['charttime']= pd.to_datetime(df_vitalsign['charttime'])

# Ομαδοποίηση καταγραφών ανά στήλη στο 8ώρο.

grouped_multiple = (df_vitalsign.groupby(['subject_id', 'stay_id',
                                          pd.Grouper(key='charttime',
                                                     freq='8H')]).agg({'temperature': ['mean'],
                                                                       'heartrate': ['mean'],
                                                                       'resprate': ['mean'],
                                                                       'o2sat': ['mean'],
                                                                       'sbp': ['mean'],
                                                                       'dbp': ['mean']}))
grouped_multiple.columns = (['Mean_temperature',
                             'Mean_heartrate',
                             'Mean_resprate',
                             'Mean_o2sat',
                             'Mean_sbp',
                             'Mean_dbp'])
grouped_multiple = grouped_multiple.reset_index()

grouped_multiple

# Κενές τιμές στον πίνακα
grouped_multiple.isna().sum()

rslt_df=pd.merge(df_ed_patients, grouped_multiple, on=['stay_id'])

# Διαγραφή της στήλης subject_id_y
rslt_df = rslt_df.drop(['subject_id_y'], axis=1)

# Μετονομασία της στήλης subject_id_x
rslt_df = rslt_df.rename(columns={'subject_id_x': 'subject_id'})

#Εξαγωγή σε csv
rslt_df.to_csv('D:\Thesis\exports\o2_ed_mean_vitals.csv', index=False)
