==> Heart_Attack/03_01_ICU_patients.ipynb <==
"""
Σε αυτό το αρχείο χρεισιμοποιούνται οι ICD9 κωδικοί που
αναφέρονται στο εγκεφαλικό και παράγεται το αρχείο
"icd9_cerebral_infarction.csv" όπου τους επεξηγεί.
Στη συνέχεια παράγεται το αρχείο "diagnoses_cerebral_infarction.csv"
όπου περιλαμβάνονται σε αυτό οι ασθενείς με εγκεφαλικό.
Έπειτα δημιουργείται το αρχείο "death.csv" όπου
εμπεριέχονται σε αυτό οι ασθενείς που απεβίωσαν.

523.740 -> Εισαγωγές στις νοσοκομειακές μονάδες.
 12.015 -> Εισαγωγές στις νοσοκομειακές μονάδες ασθενών με εγκεφαλικό.
  9.926 -> Μοναδικές εισαγωγές στις νοσοκομειακές μονάδες χωρίς περιλαμβάνονται οι ενδονοσοκομειακές μετακινήσεις.
  8.320 -> Μοναδικών ασθενών στις νοσοκομειακές μονάδες.
  7.431 -> Επιβίωσαν στις νοσοκομειακές μονάδες (89,31%).
    889 -> Απεβίωσαν στις νοσοκομειακές μονάδες (10,69%).
  
  
  5.436 -> Εισαγωγές στο σύνολο ασθενών στις ΜΕΘ.
  3.809 -> Μοναδικές εισαγωγές ασθενών στις ΜΕΘ
  3.646 -> Μοναδικοί ασθενείς στις ΜΕΘ
  2.887 -> Επιβίωσαν στις ΜΕΘ (73,71%)
    759 -> Απεβίωσαν στις ΜΕΘ (26,29%)
"""

import pandas as pd

# Δήλωση των data frame.
df_hosp_patients = pd.read_csv('D:\Thesis\exports\o3_hosp_patients.csv')
df_icd = pd.read_csv('D:\Thesis\mimic-iv-1.0\hosp\d_icd_diagnoses.csv.gz')
df_icu_stay = pd.read_csv('D:\Thesis\mimic-iv-1.0\icu\icustays.csv.gz',
                          usecols = ['subject_id', 'hadm_id'])

# 12.015 -> Εισαγωγές στις νοσοκομειακές μονάδες ασθενών με εγκεφαλικό.
df_hosp_patients.info()

# 9.926 -> Μοναδικές εισαγωγές στις νοσοκομειακές μονάδες χωρίς περιλαμβάνονται οι ενδονοσοκομειακές μετακινήσεις.
df_hosp_patients.drop_duplicates(subset=['hadm_id']).info()

# Συνολικές εισαγωγές στις ΜΕΘ (5.436)
rslt_df=pd.merge(df_hosp_patients, df_icu_stay, on=['subject_id','hadm_id'])
rslt_df.to_csv('D:\Thesis\exports\o4_icu_total_admissions.csv', index = False)

rslt_df.info()

# Μοναδικ΄ές εισαγωγές στις ΜΕΘ (3809)
rslt_df.drop_duplicates(subset=['hadm_id']).to_csv('D:\Thesis\exports\o5_icu_unique_admissions.csv', index = False)

# Μοναδικοί ασθενείς στις ΜΕΘ (3646)
rslt_df.drop_duplicates(subset=['subject_id']).info()
