==> Heart_Attack/03_03_ICU_clear_chartevent.ipynb <==
"""
Πρέπει να απομονώσω τις γραμμές με NaN από την
στήλη valuenum μιας και χρειάζομαι μόνο τις
αριθμητικές καταγραφές και όχι τα κείμενα.

26.600.037 -> Συνολικά οι εγγραφές ασθενών με
              εγκεφαλικό στις ΜΕΘ.

 9.874.474 -> Οι εγγραφές που περιέχουν τιμές.

       707 -> Αναγνωριστικοί κωδικοί μέτρησης.
"""

import pandas as pd

# Εισαγωγή dataframe
df_patient_chart = pd.read_csv('D:\Thesis\exports\o6_patient_chartevent.csv')

# Κρατώ μόνο τις γραμμές που περιέχουν τιμές στην valuenum
without_nan = df_patient_chart[df_patient_chart['valuenum'].notna()]

without_nan.to_csv('D:\Thesis\exports\o7_patient_chartevent_without_nan.csv', index=False)

# Καταμέτρηση των εξετάσεων που κάνουν οι ασθενείς με εγκεφαλικό και μετατροπή από series σε dataframe.
rslt_df = pd.DataFrame(without_nan['itemid'].value_counts(dropna=True))

# Το παραγόμενο αρχείο έχει σαν index την στήλη που περιέχει τον κωδικό των εξετάσεων, οπότε την κάνουμε reset
rslt_df = rslt_df.reset_index()

# Ορίζουμε νέες ονομασίες στις στήλες για να μπορούν να είναι ευανάγνωστες.
rslt_df = rslt_df.rename(columns={'index': 'itemid', 'itemid': 'times'})

# Εξαγωγή σε αρχείο
rslt_df.to_csv('D:\Thesis\exports\o8_chartevents_counts.csv', index=False)
