==> Heart_Attack/03_04_ICU_mean_chartevent.ipynb <==
"""
Το chartevent αποτελείται από 9.874.474
εγγραφές που αφορούν τους ασθενείς με εγκεφαλικό

Οι καταγραφές αυτές ομαδοποιήθηκαν και υπολογίσαμε
τον μέσο όρο για την κάθε μια από αυτές μέσα στο 8ώρο
όπου και ανήλθαν στα 4.177.458 εγγραφές 
"""

import pandas as pd

df_patients_chart = pd.read_csv('D:\Thesis\exports\o7_patient_chartevent_without_nan.csv')

# Μετατροπή της στήλης charttime από object σε datetime

df_patients_chart['charttime']= pd.to_datetime(df_patients_chart['charttime'])

# Ομαδοποίηση στοιχείων ανά οχτάωρο
grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                               'hadm_id',
                                               'itemid',
                                               pd.Grouper(key='charttime',
                                                          freq='8H')]).agg({'valuenum': ['mean']}))
grouped_multiple.columns = ['Mean_Chart']
grouped_multiple = grouped_multiple.reset_index()

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
grouped_multiple['date'] = pd.to_datetime(grouped_multiple['charttime']).dt.date
grouped_multiple['time'] = pd.to_datetime(grouped_multiple['charttime']).dt.time

grouped_multiple

grouped_multiple.to_csv('D:\Thesis\exports\o9_chartevent_8_hour_mean.csv', index=False)
