==> Heart_Attack/03_06_ICU_fill_chart_table.ipynb <==
"""
Γεμίζω τον πίνακα με τις μετρήσεις
του chartevet 8ώρου.

Ο πίνακας μου απαρτίζεται από 707
στήλες και 65189 γραμμές, οπότε μου
δημιουργούν 46.088.623 κελιά.

Οι τιμές μου ανέρχονται στα 4.177.458
οπότε με μια αφαίρεση προκείπτουν
41.911.165 NaN κελιά
"""

import pandas as pd

df_chart_tbl = pd.read_csv('D:\Thesis\exports\o10_empty_chart_tbl.csv', index_col=False)
df_mean_chart = pd.read_csv('D:\Thesis\exports\o9_chartevent_8_hour_mean.csv', index_col=False)
df_data_chart = pd.read_csv('D:\Thesis\mimic-iv-1.0\icu\d_items.csv.gz')

df_chart_tbl

# Φέρνω τις τελευταίες στήλες, πρώτες
for i in range (1, 5):
    cols = list(df_chart_tbl.columns)
    cols = [cols[-1]] + cols[:-1]
    df_chart_tbl = df_chart_tbl[cols]

# Ελέγχω αν υπάρχει NaN στην στήλη με τις μέσες τιμές, το αποτέλεσμα ισούται με μηδέν.
df_mean_chart['Mean_Chart'].isna().sum()

"""
Δημιουργώ εκ νέου την στήλη concat
χωρίς κενά ανάμεσα στα στοιχεία.
Θα χρησιμοποιήσω αυτή τη στήλη σαν
index γιατί τα κενά μπορεί να
δημιουργήσουν πρόβλημα.
"""
df_chart_tbl["concat"] = (df_chart_tbl["subject_id"].astype(str)
                          + df_chart_tbl["hadm_id"].astype(str)
                          + df_chart_tbl["date"].astype(str)
                          + df_chart_tbl["time"].astype(str))

"""
Αφαιρώ τους χαρακτήρες - και :
από της τιμές της concat για να
αποφύγω πιθανά προβλήματα
"""
df_chart_tbl['concat'] = df_chart_tbl['concat'].str.replace('-','')
df_chart_tbl['concat'] = df_chart_tbl['concat'].str.replace(':','')

# Ορίζω την στήλη concat ως index των γραμμών
df_chart_tbl = df_chart_tbl.set_index('concat')

"""
Δημιουργώ την στήλη concat και στον πίνακα
που περιλαμβάνει τις μετρήσεις για να είναι
το κοινό σημείο ανάμεσα στους δυο πίνακες
"""
df_mean_chart["concat"] = (df_mean_chart["subject_id"].astype(str)
                           + df_mean_chart["hadm_id"].astype(str)
                           + df_mean_chart["date"].astype(str)
                           + df_mean_chart["time"].astype(str))


"""
Αφαιρώ τους χαρακτήρες - και :
από της τιμές της concat για να
αποφύγω πιθανά προβλήματα
"""
df_mean_chart['concat'] = df_mean_chart['concat'].str.replace('-','')
df_mean_chart['concat'] = df_mean_chart['concat'].str.replace(':','')

# Γεμίζω τον πίνακα
for row in df_mean_chart.itertuples():
    con = row.concat
    item = row.itemid
    value = row.Mean_Chart
    df_chart_tbl.at['{}'.format(con), '{}'.format(item)] = value

# Κενά κελιά
df_chart_tbl.isna().sum().sum()

# Αποθήκευση του τελικού αρχείου με επικεφαλίδες τους κωδικούς των παρατηρήσεων
df_chart_tbl.to_csv('D:\Thesis\exports\o11_fill_chart_itemid_header_tbl.csv', index=False)

df_data_chart = df_data_chart.set_index('itemid')

df_data_chart

df_chart_tbl

# Ενοποίηση τιμών σε κελιά
df_data_chart["header"] = (df_data_chart["label"].astype(str)
                           + " ("
                           + df_data_chart["unitname"].astype(str)
                           + ")")

"""
Ελέγχω τις κεφαλίδες και αν δεν ανήκουν στις
τέσσερις πρώτες τιμές κάνω τις αντικατάσταση.
"""
for col in df_chart_tbl.columns:
    if ((col != 'subject_id') and
        (col != 'hadm_id') and
        (col != 'date') and
        (col != 'time')):
        temp = (df_data_chart.loc[int(col), 'header'])
        df_chart_tbl.rename(columns={'{}'.format(col): '{}'.format(temp)}, inplace=True)

# Αποθήκευση του τελικού αρχείου με επικεφαλίδες τις ονομασίες των παρατηρήσεων
df_chart_tbl.to_csv('D:\Thesis\exports\o12_fill_chart_label_header_tbl.csv', index=False)
