==> Heart_Attack/03_11_ICU_Merge_Lab_Chart.ipynb <==
"""
Ενοποίηση και των δύο πινάκων που αφορούν την
ICU σε ένα ενιαίο.

ChartEvent -> 65.189 γραμμές / 707 στήλες
LabEvent -> 68.191 γραμμές / 388 στήλες



final_table -> 90.817 γραμμές / 1095 στήλες

99.444.615 -> Θέσεις στο πίνακα.
93.990.301 -> Κενές θέσεις στον πίνακα.
 5.454.314 -> Συνολικές εγγραφές.

"""

import pandas as pd

df_icu_chart = pd.read_csv('D:\Thesis\exports\o12_fill_chart_label_header_tbl.csv')

df_lab_chart = pd.read_csv('D:\Thesis\exports\o19_fill_lab_label_header_tbl.csv')

#Ελέγχω αν μετά την συγχώνευση υπάρχουν διπλές ίδιες στήλες

rslt=pd.merge(df_icu_chart, df_lab_chart, on=['subject_id','hadm_id', 'date', 'time'],
              how="outer", suffixes=('', '_delme'))

# Και αν υπάρχουν τις αφαιρώ
rslt = rslt[[c for c in rslt.columns if not c.endswith('_delme')]]

#Εισάγω στήλη με την τιμές την ονομασία του τμήματος
rslt.insert(0, "Department", "ICU")

# Ελέγχω αν υπάρχει NaN σε ολόκληρο το dataframe
rslt.isna().sum().sum()

#Εξαγωγή σε csv
rslt.to_csv('D:\Thesis\exports\o20_ICU_merged_chart_lab.csv', index=False)
