==> Heart_Attack/03_13_Train_Test_Set.ipynb <==
## Δημιουργία Train - Test Set (4 αρχεία)

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

df_icu = pd.read_csv('D:\Thesis\exports\o31_ICU_final.csv')

y = df_icu.hospital_expire_flag

x = df_icu.drop('hospital_expire_flag',axis=1)

x_train,x_test,y_train,y_test=train_test_split(x,y,test_size=0.3)

print("shape of original dataset :", df_icu.shape)
print("shape of input - training set", x_train.shape)
print("shape of output - training set", y_train.shape)
print("shape of input - testing set", x_test.shape)
print("shape of output - testing set", y_test.shape)

#Εξαγωγή σε csv
#Χαλάει την ταξινόμηση
x_train.to_csv('D:\Thesis\exports\Train_Test_Sets\Four_files\o1_xTrain.csv', index=False)
x_test.to_csv('D:\Thesis\exports\Train_Test_Sets\Four_files\o2_xTest.csv', index=False)
y_train.to_csv('D:\Thesis\exports\Train_Test_Sets\Four_files\o3_yTrain.csv', index=False)
y_test.to_csv('D:\Thesis\exports\Train_Test_Sets\Four_files\o1_yTest.csv', index=False)

## Δημιουργία Training - Test Set (2 αρχεία)

df_icu = pd.read_csv('D:\Thesis\exports\o31_ICU_final.csv')

x_train,x_test=train_test_split(df_icu,test_size=0.3)

print("shape of original dataset :", df_icu.shape)
print("shape of input - training set", x_train.shape)

#Εξαγωγή σε csv
#Χαλάει την ταξινόμηση
x_train.to_csv('D:\Thesis\exports\Train_Test_Sets\o1_xTrain.csv', index=False)
x_test.to_csv('D:\Thesis\exports\Train_Test_Sets\o2_xTest.csv', index=False)

# Από εδώ και κάτω είναι για την παραγωγή του CSV της Βίκυς

temp=df_icu.head(63568)

temp

# εκτυπώνει όλες τις τιμές από τα dataframe.
#pd.set_option('display.max_rows', None)

temp.to_csv('D:\Thesis\exports\Train_Test_Sets\Manual_files\Test30.csv', index=False)

temp = pd.read_csv('D:\Thesis\exports\Train_Test_Sets\Manual_files\Test30.csv')

eight_hours = pd.read_csv('D:\Thesis\exports\Train_Test_Sets\Manual_files\eight_hours_tbl.csv')

rslt_df=pd.merge(temp, eight_hours, left_index=True, right_index=True)

# Φέρνω τις τελευταίες στήλες, πρώτες
cols = list(rslt_df.columns)
cols = [cols[-1]] + cols[:-1]
rslt_df = rslt_df[cols]

tempII = rslt_df.loc[rslt_df['eight_hours'] == 1]

Final = tempII.drop_duplicates(subset=['subject_id'])

Final.to_csv('D:\Thesis\exports\Train_Test_Sets\Manual_files\eight_hours_unique.csv', index=False)
