==> Heart_Attack/03_14_ICU_export_first_four_columns.ipynb <==
import pandas as pd
import numpy as np

df_icu = pd.read_csv('D:\Thesis\exports\o31_ICU_final.csv', usecols = ['subject_id', 'hadm_id', 'date', 'time'])

df_icu.to_csv('D:\Thesis\exports\o35_ICU_fist_column_data.csv', index=False)

df_icu = df_icu.drop(['subject_id', 'hadm_id', 'date', 'time', 'Department'], axis=1)

df_icu.to_csv('D:\Thesis\exports\o32_ICU_final_without_subject.csv', index=False)

df1 = df_icu.iloc[:107568]
df2 = df_icu.iloc[107568:]

df1

df_temp1 = df_icu[:107568]
df_temp2 = df_icu[-46085:]

df1 = df1.drop(['subject_id', 'hadm_id', 'date', 'time', 'Department'], axis=1)

df2 = df2.drop(['subject_id', 'hadm_id', 'date', 'time', 'Department'], axis=1)

df1.to_csv('D:\Thesis\exports\TrainII_70.csv', index=False)
df2.to_csv('D:\Thesis\exports\TestII_30.csv', index=False)
