==> Heart_Attack/03_15_Lab_Chart_by_eight_hours.ipynb <==
# Διαχωρισμός γραμμών ανά οχτάωρο

import pandas as pd
import numpy as np

df_test = pd.read_csv('D:\Thesis\Weka\Train_Test_Sets\Sorted Set\Test_30.csv')

# εκτυπώνει όλες τις τιμές από τα dataframe.
#pd.set_option('display.max_rows', None)

eight_hours = pd.read_csv('D:\Thesis\exports\Train_Test_Sets\Manual_files\eight_hours_tbl.csv')

rslt_df=pd.merge(df_test, eight_hours, left_index=True, right_index=True)

# Φέρνω τις τελευταίες στήλες, πρώτες
cols = list(rslt_df.columns)
cols = [cols[-1]] + cols[:-1]
rslt_df = rslt_df[cols]

"""
Ορίζω την τιμή του οχταώρου.
Αν θέλω να υπολογίσω την πρώτη μέρα βάζω την τιμή να είναι
μικρότερη από 4 έτσι ώστε να πιάσω τα τρία πρώτα οχτάωρα,
δηλαδή το πρώτο 24ώρο. Αν θέλω την δεύτερη μέρα μικρότερο
από 7 και ούτω καθεξής. Αν θέλω στοχευμένα οχτάωρα ορίζω
αυτό που θέλω με ισότητα.

"""
tempII = rslt_df.loc[rslt_df['eight_hours'] < 7]

#grouped_df = tempII.groupby('subject_id')



# Ομαδοποίηση στοιχείων ανά οχτάωρο
grouped_multiple = (tempII.groupby(['subject_id','hadm_id'])).mean()


df_death= pd.read_csv('D:\Thesis\mimic-iv-1.0\core\Admissions.csv.gz', usecols = ['subject_id', 'hadm_id', 'hospital_expire_flag'])

df_death = df_death.loc[df_death['hospital_expire_flag'] == 1]

# Ενοποιώ τα dataframes του συνολικού πίνακα με της θνησιμότητας.

df_rslt=pd.merge(grouped_multiple, df_death, on=['subject_id','hadm_id'],
              how="left", suffixes=('', '_delme'))

# Εισάγω την τιμή μηδέν στα κενά κελιά της hospital_expire_flag
df_rslt[['hospital_expire_flag']] = df_rslt[['hospital_expire_flag']].fillna(0)

# Μετατρέπω την στήλη hospital_expire_flag σε ακέραιο.
df_rslt['hospital_expire_flag'] = df_rslt['hospital_expire_flag'].apply(int)

"""
Αντικατάσταση τιμών από μηδέν σε
survival και από ένα σε death
στην στήλη hospital_expire_flag
"""

#df_rslt = df_rslt["hospital_expire_flag"].replace({0 : "survival", 1 :"death"}, inplace = True)


df_rslt["hospital_expire_flag"] = np.where(df_rslt["hospital_expire_flag"] == 0, "survival", "death")

df_rslt = df_rslt.drop_duplicates(subset=['subject_id'])

#Εξαγωγή σε csv
df_rslt.to_csv('D:\Thesis\exports\Chart_Lab_by_hours\First_forty_eight_hours.csv', index=False)
