==> Heart_Attack/04_01_HOSP_seperate_patients.ipynb <==
"""
Διαχωρίζω από το σύνολο των ασθενών
αυτούς που δεν εισήχθησαν στις ΜΕΘ.
"""

import pandas as pd

#Διάβασμα από csv
df_sum_hosp_admissions = pd.read_csv('D:\Thesis\exports\o3_hosp_patients.csv', usecols = ['subject_id', 'hadm_id'])

df_unique_icu = pd.read_csv('D:\Thesis\exports\o5_icu_unique_admissions.csv', usecols = ['subject_id', 'hadm_id'])

df_sum_hosp_admissions = df_sum_hosp_admissions.drop_duplicates(subset=['hadm_id'])

"""
Εισάγω μια στήλη με περιεχόμενο 1
έτσι ώστε να μπορέσω να διαχωρήσω τους ασθενείς
από τις ΜΕΘ από το σύνολο των ασθενών.
"""
df_unique_icu.insert(2, "Flag", 1)

# Ενωποιώ τους δύο πίνακες

rslt=pd.merge(df_sum_hosp_admissions, df_unique_icu, on=['subject_id','hadm_id'],
              how="outer")

# Κρατώ μόνο αυτούς που δεν έχουν τιμή στη στήλη flag
temp = rslt.loc[rslt['Flag'].isnull()]

temp = temp.drop(['Flag'], axis = 1)

temp.to_csv("D:\Thesis\exports\o21_only_hosp_patients.csv")
