==> Heart_Attack/04_03_HOSP_used_labevent.ipynb <==
"""
Κάνουμε καταμέτρηση των πιο
συχνά εμφανιζόμενων εξετάσεων.

Στο σύνολο τους 340 εγγραφές
"""

import pandas as pd

# Δήλωση των dataframes

df_labevents = pd.read_csv('D:\Thesis\exports\o23_hosp_labevents_grouped_8hours.csv')

# Καταμέτρηση των εξετάσεων που κάνουν οι ασθενείς και μετατροπή από series σε dataframe.
rslt_df = pd.DataFrame(df_labevents['itemid'].value_counts(dropna=True))

# Το παραγώμενο αρχείο έχει σαν index την στήλη που περιέχει τον κωδικό των εξετάσεων, οπότε την κάνουμε reset
rslt_df = rslt_df.reset_index()

# Και ορίζουμε νέες ονομασίες στις στήλες για να μπορούν να είναι ευανάγνωστες.
rslt_df.rename(columns={'index': 'itemid', 'itemid': 'times'}).to_csv('D:\Thesis\exports\o24_hosp_labevents_counts.csv',
                                                                      index=False)
