==> Heart_Attack/04_04_HOSP_create_lab_table.ipynb <==
"""
Δημιουργώ τον πίνακα που θα περιλαμβάνει
τις τιμές των εργαστηριακών μετρήσεων
ασθενών με εγκεφαλικό που έχουν εισαχθεί
στα νοσοκομεία αλλά όχι στις ΜΕΘ.
"""

import pandas as pd

df_rows = pd.read_csv('D:\Thesis\exports\o23_hosp_labevents_grouped_8hours.csv')

# Μετατροπή της στήλης charttime από object σε datetime
df_rows['charttime']= pd.to_datetime(df_rows['charttime'])

#Διαχωρισμός σε δυο καινούργιες στήλες της ημερομηνίας και ώρας
df_rows['date'] = pd.to_datetime(df_rows['charttime']).dt.date
df_rows['time'] = pd.to_datetime(df_rows['charttime']).dt.time

"""
Αποθηκεύω το αρχείο που πλεον περιλαμβάνει
τις δυο καινούργιες στήλες,date και time.
"""
df_rows.to_csv('D:\Thesis\exports\o25_hosp_labevents_grouped_8hours_date_time.csv',
               index=False)

list(set(df_rows['time']))

#Δημιουργώ τις στήλες από το αρχείο που περιλαμβάνει τις μοναδικές εξετάσεις
df_columns = pd.read_csv('D:\Thesis\exports\o24_hosp_labevents_counts.csv',
                         usecols = ['itemid'], index_col=False)

#Ορίζω το header ως index
df_columns = df_columns.set_index('itemid')

# Μεταφέρω τις γραμμές σε στήλες
df_columns = df_columns.transpose()

"""
Δημιουργώ ένα μοναδικό αναγνωριστικό με την ένωση των
στοιχείων από στήλες subject, hadm_id, date και time.
Ανάμεσα στα στοιχεία προβάλω ένα κενό το οποίο θα μου
φανεί χρήσιμο αργότερα για τον διαχωρισμό τους ξανά σε
διαφορετικές στήλες.
"""
df_rows["concat"] = (df_rows["subject_id"].astype(str)
                     +" "
                     +df_rows["hadm_id"].astype(str)
                     +" "
                     +df_rows["date"].astype(str)
                     +" "+
                     df_rows["time"].astype(str))

# Ελέγχω αν υπάρχει κενό κελί στην στήλη concat.
df_rows['concat'].isnull().values.any()

# Από το dr_rows κρατάω τις μοναδικές εγγραφές
df_copy_rows = pd.DataFrame(df_rows.concat.unique())

# Δημιουργώ καινούργιο πίνακα ενσωματώνοντας στον πίνακα df columns τις στήλες με τις κενές για την ώρα εξετάσεις
lab_tbl = df_columns.append(df_copy_rows)

# Μετονομάζω την στήλη 0 σε concat για να είναι ίδιες στους δυο πίνακες
lab_tbl = lab_tbl.rename({0:'concat'}, axis=1)

#Διαχωρισμός του concat σε τέσσερις στήλες
lab_tbl[['subject_id','hadm_id', 'date', 'time']] = lab_tbl['concat'].str.split(expand=True)

#Διαγράφω την στήλη concat γιατί θα πρέπει να την ξαναδημιουργήσω
#χωρίς κενά ανάμεσα στα ενοποιημένα στοιχεία, οπότε θα την έχω διπλή.
lab_tbl = lab_tbl.drop('concat', 1)

# Τρέχω το αποτέλεσμα του πίνακα
lab_tbl

lab_tbl.to_csv('D:\Thesis\exports\o26_hosp_empty_lab_tbl.csv', index=False)
