==> Heart_Attack/04_06_HOSP_ICU_Merged.ipynb <==
"""
Ενοποίηση και των δύο πινάκων που αφορούν την
ICU και HOSP.

ICU -> 90.817 γραμμές / 995 στήλες
HOSP -> 34.341 γραμμές / 340 στήλες



final_table -> 125.128 γραμμές / 1103 στήλες

 -> Θέσεις στο πίνακα.
 -> Κενές θέσεις στον πίνακα.
  -> Συνολικές εγγραφές.

"""

import pandas as pd

df_icu = pd.read_csv('D:\Thesis\exports\o20_ICU_merged_chart_lab.csv')

df_lab = pd.read_csv('D:\Thesis\exports\o28_hosp_fill_lab_label_header_tbl.csv')

#Ελέγχω αν μετά την συγχώνευση υπάρχουν διπλές ίδιες στήλες

rslt=pd.merge(df_icu, df_lab, on=['subject_id','hadm_id', 'date', 'time'],
              how="outer", suffixes=('', '_delme'))

# Και αν υπάρχουν τις αφαιρώ
rslt = rslt[[c for c in rslt.columns if not c.endswith('_delme')]]

# Ελέγχω αν υπάρχει NaN σε ολόκληρο το dataframe
rslt.isna().sum().sum()

rslt[['Department']] = rslt[['Department']].fillna('HOSP')

#Εξαγωγή σε csv
rslt.to_csv('D:\Thesis\exports\o29_HOSP_ICU_merged.csv', index=False)
